Add files via upload
Graph Neural Networks (GNNs) have fundamentally transformed the way in which we handle and examine data originating from non-Euclidean domains. Traditional approaches to imbalanced node classification problems, such as resampling, are ineffective because they do not take into account the underlying network structure of the edges. The limited methods available to capture the intricate connections encoded in the edges of a graph pose a significant challenge for GNNs in accurately classifying nodes. We propose EATSA model to enhance GNN node classification using Edge-Aware and Two-Stage Attention Mechanisms (EATSA). EATSA focuses its initial attention on edge traits, enabling the model to differentiate the variable significance of different connections between nodes, referred to as Teacher-Attention (TA). In the second step, attention is directed towards the nodes, incorporating the knowledge obtained from the edge-level analysis referred to as Student-Attention (SA). Employing this dual strategy ensures a more sophisticated comprehension of the graph's framework, resulting in improved classification precision. The EATSA model's contribution to the field of GNNs lies in its ability to utilise both node and edge information in a cohesive manner, resulting in more accurate node classifications. This highlights the essence of the model and its potential. Comparing the EATSA model to state-of-the-arts methods with two different variants shows how strong it is and how well it can handle complex problems for node classification. This solidifies its position as one of leading solution in the field of GNN architectures and their use in complex networked systems. The exceptional performance of EATSA not only showcases its effectiveness but also underscores its potential to greatly influence the future advancement of the GNN framework.